Projectile Verticality, Centering and Friendly Fire Prevention
The projectile now takes into account the y axis, the projectile spawn location has been changed so that the projectile is more in line with the crosshair. Futhermore, when the projectile spawns in it touches the player so the colliding has been altered so that it ignores the player preventing friendly fire and the destruction of the projectile.

Co-Authored-By: labcrafter <56223980+labcrafter@users.noreply.github.com>